// packages/common/src/enums/api-methods.ts
export enum ApiMethods {
  Login = 'auth/login',
  Register = 'auth/register',
  Refresh = 'auth/refresh',
  Logout = 'auth/logout',
  Categories = 'categories',
  Reviews = 'reviews',
  Orders = 'orders',
  Products = 'products',
  Statistics = 'statistics',
  Users = 'users',
  Pagination = 'pagination',
  Cart = 'cart',
}